# src/main/java/com/league/football/api/FootballApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.10).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.league.football.api;

import java.util.List;

import javax.validation.Valid;
import javax.validation.constraints.NotNull;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import com.league.football.models.StandingsTeamDetails;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-11-29T05:53:15.206Z")

@Api(value = "football", description = "the football API")
public interface FootballApi {

    @ApiOperation(value = "Get Team Standing position", nickname = "getTeamStandingPosition", notes = "", response = StandingsTeamDetails.class, responseContainer = "List", tags={ "Team Standing position", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = StandingsTeamDetails.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Team not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/football",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<StandingsTeamDetails>> getTeamStandingPosition(
    		@NotNull @ApiParam(value = "Action name", required = true, allowableValues = "get_standings, get_players, get_teams") @Valid @RequestParam(value = "action", required = true) String action,
    		@ApiParam(value = "Team name", required = false) @Valid @RequestParam(value = "teamName", required = false) String teamName,
    		@ApiParam(value = "League Name", required = false) @Valid @RequestParam(value = "leagueName", required = false) String leagueName,
    		@ApiParam(value = "Country name", required = false) @Valid @RequestParam(value = "countryName", required = false) String countryName,
    		@NotNull @ApiParam(value = "API key", required = true) @Valid @RequestParam(value = "APIkey", required = true) String apIkey,
    		@ApiParam(value = "League") @Valid @RequestParam(value = "leagueId", required = false) String leagueId);

}